Create centroid
The hardest part of a centroid extraction in R is dealing with the sign vector used to maximize positive manifold before each factor is extracted. This function implements a new algorithm from the IFI-Zurich report in references list. This will not be a function called by the user, but will be wrapped in the centroid extraction function.